server/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pandas
openpyxl
